test/requirements.txt
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
idna==2.10
iniconfig==1.1.1
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.1.2
pytest-testinfra==6.1.0
requests==2.25.0
six==1.15.0
toml==0.10.2
urllib3==1.26.2
